FIX: Images entity의 storedFileSize 필드의 타입을 int로 변경
 - 파일의 크기는 byte단위로 저장되고 서버 자체에서 받을 수 있는 파일의 최대 용량을 50MB로 설정했기 때문에 byte로 환산하면 52,428,800byte로 int형의 최대치인 21억을 넘지 않는다. 따라서 불필요한 공간 낭비를 줄이기 위해 int로 변경한다.

+++ src/main/java/net/chimhaha/clone/domain/images/Images.java
@@ -27,14 +27,14 @@ public class Images extends BaseTimeEntity {
     private String storedFilePath;
 
     @Column(nullable = false)
-    private Long storedFileSize;
+    private int storedFileSize;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
     private Posts post; // null이면 썸네일, null이 아니면 게시글 첨부 이미지
 
     @Builder
-    public Images(String realFileName, String storedFileName, String storedFilePath, Long storedFileSize, Posts post) {
+    public Images(String realFileName, String storedFileName, String storedFilePath, int storedFileSize, Posts post) {
         this.realFileName = realFileName;
         this.storedFileName = storedFileName;
         this.storedFilePath = storedFilePath;
